Fix crash when no pollers on task queue. Dedupe code.
