+++ tools/cli/taskQueue.go
@@ -24,7 +24,15 @@
 
 package cli
 
-import "github.com/urfave/cli"
+import (
+	"os"
+
+	"github.com/olekukonko/tablewriter"
+	"github.com/urfave/cli"
+	enumspb "go.temporal.io/api/enums/v1"
+	taskqueuepb "go.temporal.io/api/taskqueue/v1"
+	"go.temporal.io/server/common/primitives/timestamp"
+)
 
 func newTaskQueueCommands() []cli.Command {
 	return []cli.Command{
@@ -62,4 +70,21 @@ func newTaskQueueCommands() []cli.Command {
 			},
 		},
 	}
+}
+
+func printPollerInfo(pollers []*taskqueuepb.PollerInfo, taskQueueType enumspb.TaskQueueType) {
+	table := tablewriter.NewWriter(os.Stdout)
+	table.SetBorder(false)
+	table.SetColumnSeparator("|")
+	if taskQueueType == enumspb.TASK_QUEUE_TYPE_ACTIVITY {
+		table.SetHeader([]string{"Activity Poller Identity", "Last Access Time"})
+	} else {
+		table.SetHeader([]string{"Workflow Poller Identity", "Last Access Time"})
+	}
+	table.SetHeaderLine(false)
+	table.SetHeaderColor(tableHeaderBlue, tableHeaderBlue)
+	for _, poller := range pollers {
+		table.Append([]string{poller.GetIdentity(), formatTime(timestamp.TimeValue(poller.GetLastAccessTime()), false)})
+	}
+	table.Render()
 }
